EKUCODE: cleaned up component, move helpers and constants to other files
