+++ src/components/middle/immedia/helpers.ts
@@ -1,0 +1,3 @@
+const formatRoom = (room: string) => room.replace('-', 's');
+
+export { formatRoom };
